requirements.txt
torch
fastapi
uvicorn
chromadb
sentence-transformers 
pydantic
python-multipart 
PyMuPDF
python-docx
streamlit
requests
aiofiles
PyPDF2
mammoth
transformers
